Move view specific functions from collection to view
